// packages/design-system/src/lib/types/Button.d.ts
export type ButtonType =
  | 'primaryFill'
  | 'primaryGradientFill'
  | 'primaryOutline'
  | 'neutralFill'
  | 'neutralOutlined'
  | 'shy'
  | 'reversed';
export type ButtonSize = 'xsmall' | 'small' | 'medium' | 'large';
export type ButtonShape = 'square' | 'rounded';
export type IconPosition = 'reversed' | 'iconOnly';